refactor: create core/types.ts, migrate strategy/types imports, delete old file
