+++ src/proxy/orchestrator.ts
@@ -1,7 +1,7 @@
 import type { FastifyReply, FastifyRequest } from "fastify";
 import type { TransportResult } from "./types.js";
 import { ProviderSwitchNeeded } from "./types.js";
-import type { Target, ConcurrencyOverride } from "./strategy/types.js";
+import type { Target, ConcurrencyOverride } from "../core/types.js";
 import type { ResilienceLayer, ResilienceResult, ResilienceConfig } from "./resilience.js";
 import { ResilienceLayer as ResilienceLayerClass } from "./resilience.js";
 import type { RetryRuleMatcher } from "./retry-rules.js";
