+++ src/proxy/proxy-handler.ts
@@ -18,7 +18,7 @@ import {
 import { buildUpstreamHeaders, buildUpstreamUrl } from "./proxy-core.js";
 import { ProviderSwitchNeeded } from "./types.js";
 import type { RawHeaders, TransportResult } from "./types.js";
-import type { Target } from "./strategy/types.js";
+import type { Target } from "../core/types.js";
 import { updateLogStreamContent, updateLogClientStatus } from "../db/index.js";
 import { insertRejectedLog } from "./log-helpers.js";
 import type { RetryRuleMatcher } from "./retry-rules.js";
