--- src/proxy/strategy/types.ts
@@ -1,7 +1,0 @@
-// src/proxy/strategy/types.ts — 向后兼容 re-export，迁移完毕后删除
-export type {
-  Target,
-  ResolveContext,
-  ConcurrencyOverride,
-  ResolveResult,
-} from "../../core/types.js";
